cPort Translator with LangGraph 1.13.2026/pyproject.toml
[project]
name = "cport-translator-with-langgraph-1-13-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.6",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "livekit-agents[silero,turn-detector]~=1.3",
    "livekit-plugins-elevenlabs>=1.3.12",
    "livekit-plugins-google>=1.3.12",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-openai>=1.3.12",
    "livekit-plugins-tavus>=1.3.12",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
]
